rl/plugins/VIM/openstack/vim-manager/requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

apispec==0.39.0
appdirs==1.4.3
arrow==0.15.2
attrs==19.3.0
Babel==2.7.0
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cliff==2.16.0
cmd2==0.8.9
cryptography==2.8
daiquiri==1.6.0
debtcollector==1.22.0
decorator==4.4.0
dogpile.cache==0.8.0
flasgger==0.9.1
Flask==1.1.1
idna==2.8
importlib-metadata==0.23
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
jsonpatch==1.24
jsonpointer==2.0
jsonschema==3.1.1
keystoneauth1==3.18.0
MarkupSafe==1.1.1
marshmallow==2.20.2
marshmallow-enum==1.5.1
mistune==0.8.4
more-itertools==7.2.0
msgpack==0.6.2
munch==2.3.2
netaddr==0.7.19
netifaces==0.10.9
openstacksdk==0.37.0
os-client-config==1.33.0
os-service-types==1.7.0
osc-lib==1.14.1
oslo.config==6.11.1
oslo.context==2.23.0
oslo.i18n==3.24.0
oslo.log==3.44.1
oslo.serialization==2.29.2
oslo.utils==3.41.2
pbr==5.4.3
prettytable==0.7.2
pycparser==2.19
pyinotify==0.9.6
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyperclip==1.7.0
pyrsistent==0.15.4
python-cinderclient==5.0.0
python-dateutil==2.8.0
python-glanceclient==2.17.0
python-keystoneclient==3.22.0
python-neutronclient==6.14.0
python-novaclient==16.0.0
python-openstackclient==4.0.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requestsexceptions==1.4.0
rfc3986==1.3.2
simplejson==3.16.0
six==1.12.0
stevedore==1.31.0
urllib3==1.25.6
vim-manager==0.1.3.dev1
warlock==1.3.3
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
